// cli/src/core/api/types.ts
export interface PaginatedResponse<T> {
  totalPages: number;
  totalElements: number;
  data: T[];
}

export interface PaginationParams {
  limit?: number;
  offset?: number;
}

export interface SearchParams extends PaginationParams {
  filter?: Record<string, unknown>;
  sort?: { sortBy: string[]; order: 'ASC' | 'DESC' };
}
// ...
export interface BankRule {
  resourceId: string;
  name: string;
  actionType: string;
  appliesToReconciliationAccount: {
    code: string | null;
    currencyCode: string;
    name: string;
  };
  configuration: Record<string, unknown>;
}